+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Improved project seeding to be more realistic at representing average and edge-case examples
+* Refactored Person and Participant migrations into one
 
 ### Fixed